feat(ui): move heavy-dependency components to dedicated subpath entry points (#2522)
