+++ apps/docs/src/app/(website)/[screen]/layout.client.tsx
@@ -3,7 +3,8 @@
 import {ChevronRightIcon} from '@sanity/icons/ChevronRight'
 import {CloseIcon} from '@sanity/icons/Close'
 import {MenuIcon} from '@sanity/icons/Menu'
-import {Box, Button, Breadcrumbs, Card, Flex, Text} from '@sanity/ui'
+import {Box, Button, Card, Flex, Text} from '@sanity/ui'
+import {Breadcrumbs} from '@sanity/ui/breadcrumbs'
 import {getTheme_v2} from '@sanity/ui/theme'
 import Link from 'next/link'
 import {useState} from 'react'
