+++ apps/storybook/stories/primitives/Flex.stories.tsx
@@ -1,4 +1,5 @@
-import {Card, Code, Flex, Text} from '@sanity/ui'
+import {Card, Flex, Text} from '@sanity/ui'
+import {Code} from '@sanity/ui/code'
 import type {Meta, StoryObj} from '@storybook/react-vite'
 import {styled} from 'styled-components'
