+++ apps/storybook/stories/primitives/Popover.stories.tsx
@@ -8,15 +8,13 @@ import {
   Flex,
   LayerProvider,
   Placement,
-  Popover,
-  PopoverProps,
-  PopoverUpdateCallback,
   PortalProvider,
   Stack,
   Text,
   useClickOutsideEvent,
   useLayer,
 } from '@sanity/ui'
+import {Popover, PopoverProps, PopoverUpdateCallback} from '@sanity/ui/popover'
 import {ThemeColorToneKey} from '@sanity/ui/theme'
 import type {Meta, StoryObj} from '@storybook/react-vite'
 import {useCallback, useEffect, useMemo, useRef, useState} from 'react'
