+++ packages/themer/src/tool/ThemerSidebar.tsx
@@ -2,7 +2,9 @@ import {COLOR_TINTS, ColorTintKey} from '@sanity/color'
 import {ClipboardIcon} from '@sanity/icons/Clipboard'
 import {CloseIcon} from '@sanity/icons/Close'
 import {ResetIcon} from '@sanity/icons/Reset'
-import {Box, Button, Card, Code, Flex, Grid, Stack, Text, useToast} from '@sanity/ui'
+import {Box, Button, Card, Flex, Grid, Stack, Text} from '@sanity/ui'
+import {Code} from '@sanity/ui/code'
+import {useToast} from '@sanity/ui/toast'
 import {useMemo} from 'react'
 import {registerLanguage} from 'react-refractor'
 import typescript from 'refractor/typescript'
